Flask/requirements.txt
flask
flask_cors
flask-socketio
pillow
python-dotenv
opencv-python
numpy
pyserial
requests
simpleaudio